publish command with options to publish in a panel(cluster|resources|nested)
